feat(shell): supervise the daemon and render the app library
The window now shows your actual apps, and the shell keeps the daemon
running underneath it.

Supervision adopts rather than replaces. A healthy daemon already on the
socket is left alone - it may be serving for a terminal session someone
still wants - and only a missing, unhealthy, or version-mismatched one is
restarted. Startup happens off the main thread so the window appears
immediately and renders its connecting state instead of blocking on a
process launch.

One IPC command carries the whole socket API rather than a command per
method, which would mean editing Rust every time the API grows and would
tempt the shell into interpreting payloads it has no business
understanding.

The library is built from the desktop mockup: sidebar with per-visibility
counts and filtering, app cards with tile colours derived from the slug so
they stay put between launches, and the empty-workspace state. Every colour
comes from the token file. The network level is labelled Household
throughout, with a test asserting the word "Network" never reaches a person.

Disconnected is a designed screen, not a spinner. Each cause gets its own
sentence and a restart button, because "something went wrong" leaves
someone with nothing to do.

The tray icon is now a template image, so it takes the menu bar's own
colour instead of sitting there as a coloured square.

Two bugs found by tests rather than by luck:

- The workspace watcher was a leading-edge throttle, not a debounce. It
  fired on the first event of a burst and dropped the rest, so a folder
  deleted shortly after another was added stayed in the library until some
  unrelated change came along. Now events are coalesced and the rescan runs
  once things go quiet. The e2e suite caught it; the fix also took that
  suite from 16s to 2s.
- The "renamed" badge used a prefix check, and notes-2.local does start
  with notes. It is an equality check now.

SysStatus.uptime_secs is u32 rather than u64: ts-rs maps 64-bit integers to
bigint, which JSON cannot carry, and 136 years of uptime is enough.

Adds a nine-case end-to-end suite that spawns the real binary against a
scratch workspace and drives it over the real socket and the real HTTP
port. KT_PORTS and KT_NO_MDNS let a test pin a port and stay off the
network. The socket path deliberately lives under /tmp: macOS temp
directories are long enough to blow the 104-byte sockaddr_un limit.

+++ crates/kt-daemon/src/rpc.rs
@@ -70,7 +70,7 @@ fn handle(ctx: &Context, request: &Request) -> Result<serde_json::Value, KtError
             workspace: ctx.workspace.clone(),
             serving: ctx.serving.clone(),
             app_count: ctx.library.len() as u32,
-            uptime_secs: ctx.started.elapsed().as_secs(),
+            uptime_secs: ctx.started.elapsed().as_secs() as u32,
         }),
 
         other => Err(KtError {
